requirements.txt
Faker==20.0.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
